requirements.txt
tk
Pillow
pyinstaller
pydot-ng